Clean up some
